Fix CAN FIFO dequeuing bug that generated duplicates
Releasing the current CAN message from the hardware FIFO takes time.
While the message is being released, the FIFO depth is not decremented.
This caused the ISR to repeatedly read the same message from the FIFO.
can_read now checks the RFOM bit to determine if there are actual
messages in the fifo or if it's just a previously read message that has
not yet been released.
can_read also now only uses one FIFO for simplicity, since the second
FIFO was not being configured anyways.