Improve SLCAN stability when the UART is active
slcan_output_messages now stops formatting messages where there is no
more room in the virtual CDC-ACM output buffer.
Redo the vcdc circular buffers with unmasked head and tail counters so
that it's easier to calculate the amount of space used/available.
Redo the CAN RX circular buffer with unmasked head and tail counters so
that one more message slot can be used compared to before.
Rewrite the CAN ISR to avoid starving the MCU when message processing
slows down and the RX buffer fills up. It now disables the ISR on
overflow so that the main loop always has a chance to run.